auth: Remove SAML auth web handler
Remove the SAML auth web handler as it has moved to the enterprise
repository. The rest of the SAML connector logic has been migrated in
previous commits.

This will need to be merged shortly after the corresponding change that
adds the SAML auth web handler to the enterprise repo and should contain
that version of the enterprise repo as the `e/` submodule.